Change the Timeout to Milliseconds across diego components
[#102717334]

+++ db/sqldb/sqldb_suite_test.go
@@ -34,6 +34,7 @@ var (
 	cryptor          encryption.Cryptor
 	serializer       format.Serializer
 	migrationProcess ifrit.Process
+	useSQL           bool
 )
 
 func TestSql(t *testing.T) {
@@ -43,6 +44,11 @@ func TestSql(t *testing.T) {
 }
 
 var _ = BeforeSuite(func() {
+	useSQL = os.Getenv("USE_SQL") == "true"
+	if !useSQL {
+		return
+	}
+
 	var err error
 	fakeClock = fakeclock.NewFakeClock(time.Now())
 	fakeGUIDProvider = &fakes.FakeGUIDProvider{}
@@ -92,17 +98,27 @@ var _ = BeforeSuite(func() {
 	migrationProcess = ifrit.Invoke(migrationManager)
 })
 
+var _ = BeforeEach(func() {
+	if !useSQL {
+		Skip("SQL Backend not available")
+	}
+})
+
 var _ = AfterEach(func() {
-	truncateTables(db)
-	fakeGUIDProvider.NextGUIDReturns("", nil)
+	if useSQL {
+		truncateTables(db)
+		fakeGUIDProvider.NextGUIDReturns("", nil)
+	}
 })
 
 var _ = AfterSuite(func() {
-	migrationProcess.Signal(os.Kill)
-	_, err := db.Exec(fmt.Sprintf("DROP DATABASE diego_%d", GinkgoParallelNode()))
-	Expect(err).NotTo(HaveOccurred())
+	if useSQL {
+		migrationProcess.Signal(os.Kill)
+		_, err := db.Exec(fmt.Sprintf("DROP DATABASE diego_%d", GinkgoParallelNode()))
+		Expect(err).NotTo(HaveOccurred())
 
-	Expect(db.Close()).NotTo(HaveOccurred())
+		Expect(db.Close()).NotTo(HaveOccurred())
+	}
 })
 
 func truncateTables(db *sql.DB) {
